Social Hub/app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
psycopg[binary]==3.2.6
jinja2==3.1.5
python-multipart==0.0.18
python-dotenv==1.0.1
httpx==0.28.1
google-genai==1.5.0
apscheduler==3.11.0
pillow==11.1.0
PyJWT==2.9.0
pydantic==2.10.4
cryptography==44.0.0
